unify files and file_versions delta calculation

+++ src/modules/file/file.usecase.ts
@@ -107,12 +107,7 @@ export class FileUseCases {
   }
 
   async getUserUsedStorage(user: User): Promise<number> {
-    const [filesUsage, versionsUsage] = await Promise.all([
-      this.getUserUsedStorageIncrementally(user),
-      this.getVersionsUsageIncrementally(user),
-    ]);
-
-    return (filesUsage || 0) + (versionsUsage || 0);
+    return this.getUserUsedStorageIncrementally(user);
   }
 
   async getUserUsedStorageIncrementally(user: User): Promise<number> {
@@ -126,7 +121,11 @@ export class FileUseCases {
       );
       const today = Time.dateWithTimeAdded(1, 'day', firstUsage.period);
       const deltaChangeToday =
-        await this.fileRepository.sumFileSizeDeltaFromDate(user.id, today);
+        await this.fileRepository.sumFileSizeDeltaFromDate(
+          user.id,
+          user.uuid,
+          today,
+        );
 
       return firstUsage.delta + deltaChangeToday;
     }
@@ -138,6 +137,7 @@ export class FileUseCases {
     const deltaChangeSinceLastUsage =
       await this.fileRepository.sumFileSizeDeltaFromDate(
         user.id,
+        user.uuid,
         nextDeltaStartDate,
       );
 
@@ -168,72 +168,11 @@ export class FileUseCases {
 
     const gapDelta = await this.fileRepository.sumFileSizeDeltaBetweenDates(
       user.id,
+      user.uuid,
       startDate,
       yesterdayEndOfDay,
     );
     await this.usageService.createDailyUsage(user.uuid, yesterday, gapDelta);
-  }
-
-  async getVersionsUsageIncrementally(user: User): Promise<number> {
-    const lastVersionUsage = await this.usageService.getMostRecentVersionUsage(
-      user.uuid,
-    );
-    const calculateFirstDelta = !lastVersionUsage;
-
-    if (calculateFirstDelta) {
-      const firstUsage = await this.usageService.calculateFirstVersionUsage(
-        user.uuid,
-      );
-      const today = Time.dateWithTimeAdded(1, 'day', firstUsage.period);
-      const deltaChangeToday =
-        await this.fileVersionRepository.sumVersionSizeDeltaFromDate(
-          user.uuid,
-          today,
-        );
-
-      return firstUsage.delta + deltaChangeToday;
-    }
-
-    const aggregatedUsage =
-      await this.usageService.calculateAggregatedVersionUsage(user.uuid);
-    const nextDeltaStartDate = lastVersionUsage.getNextPeriodStartDate();
-    const deltaChangeSinceLastUsage =
-      await this.fileVersionRepository.sumVersionSizeDeltaFromDate(
-        user.uuid,
-        nextDeltaStartDate,
-      );
-
-    const hasYesterdaysUsage = Time.isToday(nextDeltaStartDate);
-    if (!hasYesterdaysUsage) {
-      await this.backfillVersionsUsageUntilYesterday(
-        user,
-        nextDeltaStartDate,
-      ).catch((error) => {
-        new Logger('[USAGE/VERSION_BACKFILL]').error(
-          {
-            user: { email: user.email, uuid: user.uuid, userId: user.id },
-            nextDeltaStartDate,
-            error,
-          },
-          'There was an error backfilling user version usage',
-        );
-      });
-    }
-
-    return aggregatedUsage + deltaChangeSinceLastUsage;
-  }
-
-  async backfillVersionsUsageUntilYesterday(user: User, startDate: Date) {
-    const yesterday = Time.dateWithTimeAdded(-1, 'day');
-    const yesterdayEndOfDay = Time.endOfDay(yesterday);
-
-    const gapDelta =
-      await this.fileVersionRepository.sumVersionSizeDeltaBetweenDates(
-        user.uuid,
-        startDate,
-        yesterdayEndOfDay,
-      );
-    await this.usageService.createVersionUsage(user.uuid, yesterday, gapDelta);
   }
 
   async deleteFilePermanently(
@@ -963,36 +902,19 @@ export class FileUseCases {
       file.size,
     );
 
+    const { size, modificationTime } = newFileData;
+    const { fileId: oldFileId, bucket } = file;
+
     if (shouldVersion) {
       await this.applyRetentionPolicy(fileUuid, user.uuid);
-
-      const { fileId, size, modificationTime } = newFileData;
-
-      await Promise.all([
-        this.fileVersionRepository.upsert({
-          fileId: file.uuid,
-          userId: user.uuid,
-          networkFileId: file.fileId,
-          size: file.size,
-          status: FileVersionStatus.EXISTS,
-        }),
-        this.fileRepository.updateByUuidAndUserId(fileUuid, user.id, {
-          fileId: newFileId,
-          size,
-          updatedAt: new Date(),
-          ...(modificationTime ? { modificationTime } : null),
-        }),
-      ]);
-
-      return {
-        ...file.toJSON(),
-        fileId: newFileId,
-        size,
-      };
+      await this.fileVersionRepository.upsert({
+        fileId: file.uuid,
+        userId: user.uuid,
+        networkFileId: file.fileId,
+        size: file.size,
+        status: FileVersionStatus.EXISTS,
+      });
     }
-
-    const { fileId: oldFileId, bucket } = file;
-    const { size, modificationTime } = newFileData;
 
     await this.fileRepository.updateByUuidAndUserId(fileUuid, user.id, {
       fileId: newFileId,
@@ -1017,7 +939,7 @@ export class FileUseCases {
       });
     });
 
-    if (oldFileId) {
+    if (!shouldVersion && oldFileId) {
       try {
         await this.network.deleteFile(user, bucket, oldFileId);
       } catch (error) {
